feat: add auth and validations

+++ src/application/task.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable, NotFoundException } from "@nestjs/common";
+import { ForbiddenException, Inject, Injectable, NotFoundException } from "@nestjs/common";
 import { TaskRepository } from "src/domain/repositories/task.repository";
 import { Task } from "src/domain/task";
 import { v4 } from "uuid";
@@ -12,23 +12,30 @@ export class TaskService {
         return this.repository.getAllTaskByUser(id);
      }
 
-    async saveTask(task: Omit<Task, "id">): Promise<void> {
+    async saveTask(task: Omit<Task, "id" | "userId">, userId: string): Promise<void> {
         const taskCreate: Task = {
             id: v4(),
-            ...task
+            ...task,
+            userId: userId
         }
         await this.repository.save(taskCreate);
     }
 
-    async deleteTask(id: string): Promise<void> {
-        await this.getTaskByIdOrFail(id);
+    async deleteTask(id: string, userId: string): Promise<void> {
+        const task = await this.getTaskByIdOrFail(id);
+        if(task?.userId !== userId){
+            throw new ForbiddenException();
+        }
         await this.repository.delete(id);
     }
 
-    async changeStatusOk(id: string, status: boolean): Promise<void> {
+    async changeStatusOk(id: string, status: boolean, userId: string): Promise<void> {
         const task = await this.getTaskByIdOrFail(id);
         if (task?.isDone !== false) {
             throw new NotFoundException('Task status cannot be changed');
+        }
+        if(task?.userId !== userId){
+            throw new ForbiddenException();
         }
         await this.repository.updateStatus(id, status);
     }
